requirements.txt
openai-whisper
streamlit
pandas
ffmpeg-python
